refactor: all chip-oriented files should be in subdirectory (perhaps invisible for user/application)
